refactor(app): initial commit
