--- .nvmrc
@@ -1,1 +1,0 @@
-16.10.0
