+++ README.md
@@ -1,11 +1,20 @@
-# template-frontend
+# @pagopa/cra-template-typescript
 
-## To configure the workspace execute the following commands
-- yarn install
-- yarn generate
+This is the official TypeScript template for [@pagopa/cra-template-typescript](https://github.com/pagopa/cra-template-typescript).
 
-## To execute locally a configured workspace execute the following command
-- yarn start
+To use this template, add `--template @pagopa/typescript` when creating a new app.
 
-## To build a configured workspace execute the following command
-- yarn build
+For example:
+
+```sh
+npx create-react-app my-app --template @pagopa/typescript
+
+# or
+
+yarn create react-app my-app --template @pagopa/typescript
+```
+
+For more information, please refer to:
+
+- [Getting Started](https://create-react-app.dev/docs/getting-started) – How to create a new app.
+- [User Guide](https://create-react-app.dev) – How to develop apps bootstrapped with Create React App.
