+++ package.json
@@ -1,105 +1,33 @@
 {
-  "name": "template-frontend",
-  "version": "1.0.0",
-  "homepage": "/",
-  "private": true,
-  "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.3",
-    "@mui/material": "^5.0.3",
-    "@mui/x-data-grid": "^5.0.1",
-    "@mui/x-data-grid-generator": "^5.0.1",
-    "@pagopa/mui-italia": "pagopa/mui-italia",
-    "@pagopa/ts-commons": "^10.2.1",
-    "@reduxjs/toolkit": "^1.6.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/lodash": "^4.14.172",
-    "@types/node": "^12.20.19",
-    "@types/react": "^17.0.18",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/redux-logger": "^3.0.9",
-    "axios": "^0.23.0",
-    "core-js": "^3.20.2",
-    "env-var": "^7.1.1",
-    "formik": "^2.2.9",
-    "fp-ts": "^2.11.5",
-    "io-ts": "^2.2.16",
-    "lodash": "^4.17.21",
-    "node-fetch": "^3.0.0",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^3.0.0",
-    "react-dom": "^17.0.2",
-    "react-dropzone": "^11.4.2",
-    "react-md-spinner": "^1.0.0",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux-logger": "^3.0.6",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.1.2"
-  },
+  "name": "@pagopa/cra-template-typescript",
+  "version": "0.1.0",
+  "main": "template.json",
   "scripts": {
-    "clean": "npm-run-all clean:*",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "clean:build": "rimraf build",
-    "test": "react-scripts test -i",
-    "test:coverage": "react-scripts test -i --coverage",
-    "clean:coverage": "rimraf coverage",
-    "eject": "react-scripts eject",
-    "lint-autofix": "eslint . -c .eslintrc.js --ext .ts,.tsx --fix",
-    "lint": "eslint . -c .eslintrc.js --ext .ts,.tsx ",
-    "prettify": "prettier --write \"./**/*.{ts,tsx}\"",
+    "format": "prettier --write ./template/src/**/*.{ts,tsx}"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+  "keywords": [
+    "react",
+    "create-react-app",
+    "template",
+    "typescript"
+  ],
+  "description": "The base TypeScript template for PagoPA based on Create React App.",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/pagopa/cra-template-typescript"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "license": "MIT",
+  "engines": {
+    "node": ">=14"
   },
+  "bugs": {
+    "url": "https://github.com/pagopa/cra-template-typescript/issues"
+  },
+  "files": [
+    "template",
+    "template.json"
+  ],
   "devDependencies": {
-    "@pagopa/openapi-codegen-ts": "^10.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "api-spec-converter": "^2.12.0",
-    "eslint": "7.11.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-functional": "^3.7.2",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-sonarjs": "^0.10.0",
-    "mkdirp": "^1.0.4",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.4.1",
-    "regex-replace": "^2.3.1",
-    "rimraf": "^3.0.2"
-  },
-  "jest": {
-    "restoreMocks": true,
-    "clearMocks": true,
-    "resetMocks": true,
-    "coveragePathIgnorePatterns": [
-      "src/index.tsx",
-      "src/reportWebVitals.ts",
-      "src/api/generated"
-    ]
+    "prettier": "^2.5.1"
   }
 }
