--- public/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="it">
-  <head>
-    <base href="%PUBLIC_URL%/" />
-
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta http-equiv="x-ua-compatible" content="ie=edge" />
-
-    <title>Template</title>
-    <meta name="description" content="Modulo di Dashboard per la piattaforma Self Care di PagoPA" />
-    <meta name="theme-color" content="#00a1b0" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.webmanifest" crossorigin="anonymous" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon-32x32.png" type="image/png" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.svg" type="image/svg+xml" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-48x48.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-72x72.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-96x96.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-144x144.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-192x192.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-256x256.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-384x384.png" />
-    <link rel="apple-touch-icon" sizes="48x48" href="%PUBLIC_URL%/icons/icon-512x512.png" />
-
-    <link href="https://selfcare.pagopa.it/assets/font/selfhostedfonts.css" rel="stylesheet" />
-  </head>
-  <body>
-    <noscript>Abilita JavaScript per utilizzare quest'applicazione</noscript>
-    <div id="root"></div>
-  </body>
-</html>
