--- src/index.css
@@ -1,9 +1,0 @@
-body {
-  margin: 0;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.MuiTypography-root {
-  line-height: 1.5 !important;
-}
