--- src/react-app-env.d.ts
@@ -1,10 +1,0 @@
-/// <reference types="react-scripts" />
-
-declare namespace NodeJS {
-  interface ProcessEnv {
-    NODE_ENV: 'development' | 'uat' | 'production';
-  }
-}
-interface Window {
-  Stripe: any;
-}
