+++ template.json
@@ -1,0 +1,38 @@
+{
+  "package": {
+    "dependencies": {
+      "@pagopa/mui-italia": "^0.4.2",
+      "@testing-library/jest-dom": "^5.14.1",
+      "@testing-library/react": "^12.0.0",
+      "@testing-library/user-event": "^13.2.1",
+      "@types/jest": "^27.0.1",
+      "@types/node": "^16.7.13",
+      "@types/react": "^17.0.20",
+      "@types/react-dom": "^17.0.9",
+      "typescript": "^4.4.2",
+      "web-vitals": "^2.1.0"
+    },
+    "scripts": {
+      "lint": "eslint . -c .eslintrc.js --ext .ts,.tsx --fix",
+      "format": "prettier \"./**/*.{ts,tsx}\" --write"
+    },
+    "devDependencies": {
+      "@typescript-eslint/eslint-plugin": "^5.1.0",
+      "@typescript-eslint/parser": "^5.1.0",
+      "danger": "^10.5.3",
+      "danger-plugin-digitalcitizenship": "^1.0.0",
+      "eslint": "^7.11.0",
+      "eslint-plugin-functional": "^3.7.2",
+      "eslint-plugin-import": "^2.24.2",
+      "eslint-plugin-react": "^7.26.1",
+      "eslint-plugin-sonarjs": "^0.10.0",
+      "eslint-config-prettier": "^8.3.0",
+      "eslint-plugin-prettier": "^4.0.0",
+      "eslint-plugin-react-hooks": "^4.2.0",
+      "prettier": "^2.4.1"
+    },
+    "eslintConfig": {
+      "extends": ["react-app", "react-app/jest"]
+    }
+  }
+}
