+++ template/.gitignore
@@ -13,6 +13,7 @@
 
 # misc
 .DS_Store
+.env.local
 .env.development.local
 .env.test.local
 .env.production.local
