+++ template/README.md
@@ -1,0 +1,31 @@
+<p align="center">
+  <img src="./src/assets/imgs/pagopa-logo.svg" alt="pagopa-logo" width="120px" height="120px"/>
+  <br>
+  <i>cra-template-typescript</i>
+  <br>
+</p>
+
+<p align="center">
+  <a href="https://www.pagopa.it"><strong>www.pagopa.it</strong></a>
+  <br>
+</p>
+
+# Table of contents
+
+- [Getting Started](#getting-started)
+- [Scripts](#scripts)
+
+## Getting Started
+
+This project was bootstrapped with [@pagopa/cra-template-typescript](https://github.com/pagopa/cra-template-typescript).
+
+## Scripts
+
+In the project directory, you can run:
+
+- `npm run start` - runs the app in the development mode
+- `npm run test` - launches the test runner in the interactive watch mode
+- `npm run build` - builds the app for production to the `build` folder
+- `npm run eject` - eject configuration choices
+
+**[⬆ back to top](#table-of-contents)**
