+++ template/public/manifest.webmanifest
@@ -1,9 +1,9 @@
 {
-  "name": "pagopa interop",
-  "short_name": "pagopa interop",
+  "name": "pagopa template",
+  "short_name": "pagopa template",
   "start_url": "/",
   "background_color": "#FFFFFF",
-  "theme_color": "#00a1b0",
+  "theme_color": "#0073E6",
   "display": "minimal-ui",
   "icons": [
     {
