--- tsconfig.json
@@ -1,40 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "esnext",
-    "lib": [
-      "es6",
-      "dom",
-      "es2019"
-    ],
-    "jsx": "react-jsx",
-    "noEmit": true,
-    "strict": true,
-    "noImplicitAny": true,
-    "noUnusedLocals": true,
-    "noImplicitReturns": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "allowSyntheticDefaultImports": true,
-    "preserveConstEnums": true,
-    "allowJs": false,
-    "noUnusedParameters": true,
-    "strictFunctionTypes": true,
-    "isolatedModules": true,
-    "noFallthroughCasesInSwitch": true,
-    "resolveJsonModule": true
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "__mocks__",
-    "dist",
-    "node_modules",
-    "**/__tests__/*",
-    "Dangerfile.ts"
-  ]
-}
